Only publish win32 archive and installer artifacts
The Linux archive is not very useful as it's not very portable: it won't
run unless the operating environment is very similar to the build
environment.

The win64 archive is portable betwen 64-bit Windows installs, but is
only really built as a proof-of-concept: we have no plans to release it.
As such, publishing the artifact is not worth its storage usage - if we
want to test a win64 build out, we can always create a branch that does
publish it.